requirements.txt
streamlit
streamlit-code-editor
pandas
ruamel.yaml